+++ app/jobs/create_activity_notification_job.rb
@@ -1,0 +1,10 @@
+# Send notifications about creating events
+class CreateActivityNotificationJob < ApplicationJob
+  queue_as :default
+
+  def perform(activity_id:)
+    activity = Activity.find_by(id: activity_id)
+    return if activity.nil?
+    Notifies::CreateActivity.new.call(activity: activity)
+  end
+end