+++ app/models/activity.rb
@@ -1,0 +1,4 @@
+# Represents news from guilds
+class Activity < ApplicationRecord
+  belongs_to :guild
+end