+++ config/routes.rb
@@ -80,6 +80,7 @@ Rails.application.routes.draw do
       resources :events, only: %i[index] do
         get :filter_values, on: :collection
       end
+      resources :activities, only: %i[show create update]
     end
   end
 
@@ -103,6 +104,7 @@ Rails.application.routes.draw do
     end
     resources :recipes, only: %i[index new edit destroy]
     resources :guilds, only: %i[index show new edit] do
+      get :activities, on: :member
       get :management, on: :member
       get :statics, on: :member
       get :notifications, on: :member
@@ -119,6 +121,7 @@ Rails.application.routes.draw do
       get :decline, on: :member
     end
     resources :deliveries, only: %i[new destroy]
+    resources :activities, only: %i[index new edit destroy]
     resources :guild_invites, only: %i[] do
       get :find, on: :collection
     end