+++ db/seeds.rb
@@ -270,3 +270,9 @@ notification_form.persist?
 
 notification_form = NotificationForm.new(name: { 'en' => 'Bank request creation', 'ru' => 'Создание заявки в банке' }, event: 'bank_request_creation', status: 1)
 notification_form.persist?
+
+notification_form = NotificationForm.new(name: { 'en' => 'Activity creation', 'ru' => 'Создание новости' }, event: 'activity_creation', status: 0)
+notification_form.persist?
+
+notification_form = NotificationForm.new(name: { 'en' => 'Activity creation', 'ru' => 'Создание новости' }, event: 'activity_creation', status: 1)
+notification_form.persist?