+++ spec/factories/activities.rb
@@ -1,0 +1,7 @@
+FactoryBot.define do
+  factory :activity do
+    title { 'Title' }
+    description { 'Description' }
+    association :guild
+  end
+end