+++ spec/services/notifies/create_activity_spec.rb
@@ -1,10 +1,10 @@
-RSpec.describe Notifies::CreateEvent, type: :service do
+RSpec.describe Notifies::CreateActivity, type: :service do
   let!(:character) { create :character }
   let!(:guild) { create :guild, world: character.world, fraction: character.race.fraction, world_fraction: character.world_fraction }
   let!(:time_offset) { create :time_offset, timeable: guild }
-  let!(:guild_event) { create :event, eventable: guild }
-  let!(:guild_notification) { create :notification, event: 'guild_event_creation', status: 0 }
-  let!(:user_notification) { create :notification, event: 'guild_event_creation', status: 1 }
+  let!(:activity) { create :activity, guild: guild }
+  let!(:guild_notification) { create :notification, event: 'activity_creation', status: 0 }
+  let!(:user_notification) { create :notification, event: 'activity_creation', status: 1 }
   let!(:delivery1) { create :delivery, deliveriable: guild, notification: guild_notification }
   let!(:delivery_param) { create :delivery_param, delivery: delivery1 }
   let!(:delivery2) { create :delivery, deliveriable: character.user, notification: user_notification }
@@ -16,7 +16,7 @@ RSpec.describe Notifies::CreateEvent, type: :service do
     it 'calls perform delivery' do
       expect(PerformDelivery).to receive(:call).and_call_original
 
-      service.call(event: guild_event)
+      service.call(activity: activity)
     end
   end
 
@@ -24,7 +24,7 @@ RSpec.describe Notifies::CreateEvent, type: :service do
     it 'calls perform delivery' do
       expect(PerformDelivery).to receive(:call).and_call_original
 
-      service.send(:notify_guild, guild_id: guild.id, event_name: 'guild_event_creation', event: guild_event)
+      service.send(:notify_guild, guild_id: guild.id, event_name: 'activity_creation', activity: activity)
     end
   end
 
@@ -34,7 +34,7 @@ RSpec.describe Notifies::CreateEvent, type: :service do
     it 'calls perform delivery' do
       expect(PerformDelivery).to receive(:call).and_call_original
 
-      service.send(:notify_users, object: guild, event_name: 'guild_event_creation', event: guild_event)
+      service.send(:notify_users, object: guild, event_name: 'activity_creation', activity: activity)
     end
   end
 end