+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- model Activity
+- ActivityDryForm
+- Activities page for guilds
+- creating activities
+- activities page with last guild news
+- update/delete activities
+- activity's notifications
+
 ## [1.1.9] - 2019-12-24
 ### Modified
 - event calendar with subscribe statuses